дабля топчусь на месте. 0. учить вопросы =) 1. дохерачить устройства 2. читать переменные из устройства 3. читать переменные из веба 4. сделать структуру value table и читать переменные по ней 5. апдейт устройств и юзверей 6. рассылка мыла по ивентам 7. складывать переменные в архив 8. при регистрации юзеров - валидация полей

+++ webscada/api/src/main/java/webscada/api/dao/IDevJPADao.java
@@ -3,7 +3,7 @@ package webscada.api.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import webscada.entity.Dev;
 
-public interface IDevJPADao extends JpaRepository<Dev, Integer> {
+public interface IDevJPADao extends JpaRepository<Dev, Long> {
 
 	Dev findByDevName(String dev);
 }
